refactor: utils 폴더 추가 및 파일명 변경
